Refactoring - merging unnecessary toString and toInt into one inspection
