+++ README.md
@@ -139,7 +139,7 @@ For instructions on suppressing warnings by file, by inspection or by line see [
 
 ### Inspections
 
-There are currently 117 inspections. An overview list is given, followed by a more detailed description of each inspection after the list (todo: finish rest of detailed descriptions)
+There are currently 116 inspections. An overview list is given, followed by a more detailed description of each inspection after the list (todo: finish rest of detailed descriptions)
 
 | Name | Brief Description | Default Level |
 |------|-------------------|---------------|
@@ -242,10 +242,9 @@ There are currently 117 inspections. An overview list is given, followed by a mo
 | SwapSortFilter | `sort.filter` can be replaced with `filter.sort` for performance | Info |
 | TryGet | Checks for use of `Try.get` | Error |
 | TypeShadowing | Checks for shadowed type parameters in methods | Warning |
+| UnnecessaryConversion | Checks for unnecessary `toInt` on instances of Int or `toString` on Strings, etc. | Warning |
 | UnnecessaryIf | Checks for code like `if (expr) true else false` | Info |
 | UnnecessaryReturnUse | Checks for use of `return` keyword in blocks | Info |
-| UnnecessaryToInt | Checks for unnecessary `toInt` on instances of Int | Warning |
-| UnnecessaryToString | Checks for unnecessary `toString` on instances of String | Warning |
 | UnreachableCatch | Checks for catch clauses that cannot be reached | Warning |
 | UnsafeContains | Checks for `List.contains(value)` for invalid types | Error |
 | UnsafeStringContains | Checks for `String.contains(value)` for invalid types | Error |
