+++ src/main/scala/com/sksamuel/scapegoat/ScapegoatConfig.scala
@@ -122,10 +122,9 @@ object ScapegoatConfig extends App {
     new UnsafeTraversableMethods,
     new TryGet,
     new TypeShadowing,
+    new UnnecessaryConversion,
     new UnnecessaryIf,
     new UnnecessaryReturnUse,
-    new UnnecessaryToInt,
-    new UnnecessaryToString,
     new UnreachableCatch,
     new UnsafeContains,
     new UnsafeStringContains,
